Remove dependency of commons-cli
