+++ maven/IMG2PDF/pom.xml
@@ -50,12 +50,6 @@
             <scope>test</scope>
         </dependency>
 
-        <!-- https://mvnrepository.com/artifact/commons-cli/commons-cli -->
-        <dependency>
-            <groupId>commons-cli</groupId>
-            <artifactId>commons-cli</artifactId>
-            <version>1.4</version>
-        </dependency>
 
         <dependency>
             <groupId>info.picocli</groupId>
