+++ maven/IMG2PDF/src/main/java/org/vincentyeh/IMG2PDF/commandline/MainTestProgram.java
@@ -1,4 +1,4 @@
-package org.vincentyeh.IMG2PDF.newcommandline;
+package org.vincentyeh.IMG2PDF.commandline;
 
 import org.vincentyeh.IMG2PDF.SharedSpace;
 import picocli.CommandLine;
@@ -9,6 +9,5 @@ public class MainTestProgram {
         CommandLine cmd= new CommandLine(new IMG2PDFCommand());
         int exitCode =cmd.execute(args);
         System.exit(exitCode);
-
     }
 }
