--- maven/IMG2PDF/src/main/java/org/vincentyeh/IMG2PDF/commandline/action/Action.java
@@ -1,11 +1,0 @@
-package org.vincentyeh.IMG2PDF.commandline.action;
-
-/**
- * Action of command line.This interface make its subclass need to initialize and do stuff. 
- * 
- * @author vincent
- */
-public interface Action {
-	
-	void start() throws Exception;
-}
