--- maven/IMG2PDF/src/main/java/org/vincentyeh/IMG2PDF/commandline/action/ActionMode.java
@@ -1,19 +1,0 @@
-package org.vincentyeh.IMG2PDF.commandline.action;
-
-import org.vincentyeh.IMG2PDF.commandline.parser.ActionParser;
-import org.vincentyeh.IMG2PDF.commandline.parser.ConvertActionParser;
-import org.vincentyeh.IMG2PDF.commandline.parser.CreateActionParser;
-
-public enum ActionMode {
-	create(new CreateActionParser()),convert(new ConvertActionParser());
-
-	final ActionParser parser;
-
-	ActionMode(ActionParser parser) {
-		this.parser = parser;
-	}
-
-	public ActionParser getParser() {
-		return parser;
-	}
-}
