--- maven/IMG2PDF/src/main/java/org/vincentyeh/IMG2PDF/commandline/action/CreateAction.java
@@ -1,149 +1,0 @@
-package org.vincentyeh.IMG2PDF.commandline.action;
-
-import java.io.*;
-import java.nio.charset.Charset;
-import java.nio.file.Files;
-import java.util.*;
-
-import org.vincentyeh.IMG2PDF.SharedSpace;
-import org.vincentyeh.IMG2PDF.commandline.parser.core.HandledException;
-import org.vincentyeh.IMG2PDF.task.*;
-import org.vincentyeh.IMG2PDF.task.factory.DirlistTaskFactory;
-import org.vincentyeh.IMG2PDF.task.factory.exception.DirListException;
-import org.vincentyeh.IMG2PDF.task.factory.exception.SourceFileException;
-import org.vincentyeh.IMG2PDF.util.file.GlobbingFileFilter;
-import org.vincentyeh.IMG2PDF.util.file.FileSorter;
-public class CreateAction implements Action {
-    private final FileSorter fileSorter;
-    protected final GlobbingFileFilter ffh;
-
-    //    For PDF
-    private final DocumentArgument documentArgument;
-    private final PageArgument pageArgument;
-    private final String pdf_dst;
-
-    //    For tasklist
-    private final File tasklist_dst;
-    private final boolean debug;
-    private final boolean overwrite;
-    private final File[] sourceFiles;
-
-    private CreateAction(FileSorter fileSorter, GlobbingFileFilter ffh, DocumentArgument documentArgument, PageArgument pageArgument, String pdf_dst, File tasklist_dst, boolean debug, boolean overwrite, File[] sourceFiles) {
-        this.fileSorter = fileSorter;
-        this.ffh = ffh;
-        this.documentArgument = documentArgument;
-        this.pageArgument = pageArgument;
-        this.pdf_dst = pdf_dst;
-        this.tasklist_dst = tasklist_dst;
-        this.debug = debug;
-        this.overwrite = overwrite;
-        this.sourceFiles = sourceFiles;
-    }
-
-    @Override
-    public void start() throws Exception {
-
-        if (!overwrite && tasklist_dst.exists()) {
-            System.err.printf(SharedSpace.getResString("public.err.overwrite") + "\n", tasklist_dst.getAbsolutePath());
-            throw new HandledException(new RuntimeException("Overwrite deny"), getClass());
-        }
-        DirlistTaskFactory.setArgument(documentArgument, pageArgument, pdf_dst);
-        DirlistTaskFactory.setImageFilesRule(ffh, fileSorter);
-
-        List<Task> tasks = new ArrayList<>();
-        try {
-            for (File dirlist : sourceFiles) {
-                tasks.addAll(DirlistTaskFactory.createFromDirlist(dirlist, SharedSpace.Configuration.DIRLIST_READ_CHARSET));
-            }
-        }catch (SourceFileException | DirListException e){
-//            System.err.println(e.getMessage());
-            throw new HandledException(e,getClass());
-        }
-        save(tasks, tasklist_dst);
-    }
-
-    public void save(List<Task> taskList, File destination) throws HandledException {
-        destination.getParentFile().mkdirs();
-
-        String content= new TaskListConverter().toXml(taskList);
-        try {
-            writeStringToFile(destination,content,SharedSpace.Configuration.TASKlIST_WRITE_CHARSET);
-            System.out.printf("[" + SharedSpace.getResString("public.info.exported") + "] %s\n", tasklist_dst.getAbsolutePath());
-        } catch (IOException e) {
-            System.err.printf(SharedSpace.getResString("create.err.tasklist_create") + "\n", e.getMessage());
-            throw new HandledException(e, getClass());
-        }catch (Exception e){
-            e.printStackTrace();
-            throw new HandledException(e,getClass());
-        }
-    }
-
-    private void writeStringToFile(File file, String content, Charset charset) throws IOException {
-        List<String> contents=new ArrayList<>();
-        contents.add(content);
-        Files.write(file.toPath(),contents, charset);
-    }
-    public static class Builder {
-        private FileSorter fileSorter;
-        protected GlobbingFileFilter ffh;
-
-        private DocumentArgument documentArgument;
-        private PageArgument pageArgument;
-        private String pdf_dst;
-
-        private File tasklist_dst;
-        private boolean debug;
-        private boolean overwrite;
-        private File[] sourceFiles;
-
-        public Builder setFileSorter(FileSorter fileSorter) {
-            this.fileSorter = fileSorter;
-            return this;
-        }
-
-        public Builder setFileFilterHelper(GlobbingFileFilter ffh) {
-            this.ffh = ffh;
-            return this;
-        }
-
-        public Builder setDocumentArgument(DocumentArgument documentArgument) {
-            this.documentArgument = documentArgument;
-            return this;
-        }
-
-        public Builder setPageArgument(PageArgument pageArgument) {
-            this.pageArgument = pageArgument;
-            return this;
-        }
-
-        public Builder setPdfDestination(String pdf_dst) {
-            this.pdf_dst = pdf_dst;
-
-            return this;
-        }
-
-        public Builder setTasklistDestination(File tasklist_dst) {
-            this.tasklist_dst = tasklist_dst;
-            return this;
-        }
-
-        public Builder setDebug(boolean debug) {
-            this.debug = debug;
-            return this;
-        }
-
-        public Builder setOverwrite(boolean overwrite) {
-            this.overwrite = overwrite;
-            return this;
-        }
-
-        public Builder setSourceFiles(File[] sourceFiles) {
-            this.sourceFiles = sourceFiles;
-            return this;
-        }
-
-        public CreateAction build() {
-            return new CreateAction(fileSorter, ffh, documentArgument, pageArgument, pdf_dst, tasklist_dst, debug, overwrite, sourceFiles);
-        }
-    }
-}
