+++ maven/IMG2PDF/src/main/java/org/vincentyeh/IMG2PDF/commandline/converter/AccessPermissionConverter.java
@@ -1,4 +1,4 @@
-package org.vincentyeh.IMG2PDF.newcommandline.converter;
+package org.vincentyeh.IMG2PDF.commandline.converter;
 
 import org.apache.pdfbox.pdmodel.encryption.AccessPermission;
 import picocli.CommandLine;
