+++ maven/IMG2PDF/src/main/java/org/vincentyeh/IMG2PDF/commandline/converter/FileSorterConverter.java
@@ -1,4 +1,4 @@
-package org.vincentyeh.IMG2PDF.newcommandline.converter;
+package org.vincentyeh.IMG2PDF.commandline.converter;
 
 import org.vincentyeh.IMG2PDF.util.file.FileSorter;
 import picocli.CommandLine;
